Feature/vih 4036 switch on cam and mic (#37)
* request user cam and mic on demand

* import web rtc adapter to support older browsers

* change navigator to any to resolve compilation errors

* hide and show 'Watch Video' button based on whether or not mic and camera were granted access

* fix typo in message for switching on camera and mic

* update stylesheets to match the GOV UK Template

* remove use of <br /> to be more screen reader friendly

* go to video fix

* Added acceptance tests

* Corrected error in feature file

* fixed locator change

* Added page

* Fixed page journey

* Fixed tests

* Fixed timeout issue

* direct to camera question first

* Fixed waiting room page objects

* Fixing more tests - equipment checks

* Fixed more tests - Waiting room and Switch on

* replace div with p for information tag when no is selected

* Updated Bookings api client

* Added asserts for test settings

* Removing the dependency on bookings-api for now until a better solution is found

+++ VideoWeb/VideoWeb.AcceptanceTests/Hooks/DataSetupHooks.cs
@@ -50,7 +50,7 @@ namespace VideoWeb.AcceptanceTests.Hooks
         [AfterScenario]
         public static void RemoveHearing(TestContext context, HearingsEndpoints endpoints)
         {
-            if (context.NewHearingId == Guid.Empty || context.NewHearingId == null) return;
+            if (context.NewHearingId == Guid.Empty || context.NewHearingId == null || context.HearingIsNotInBookingsDb) return;
             context.Request = context.Delete(endpoints.RemoveHearing(context.NewHearingId));
             context.Response = context.BookingsApiClient().Execute(context.Request);
             context.Response.IsSuccessful.Should().BeTrue("New hearing has been deleted after the test");
